pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioformer"
version = "0.1.0"
description = "Scaffold for a bioreactor transformer proof of concept."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "ipykernel>=6.29.5",
  "jupyterlab>=4.4.1",
  "matplotlib>=3.10.1",
  "numpy>=2.2.4",
  "pandas>=2.2.3",
  "pyarrow>=19.0.1",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.6.1",
  "scipy>=1.15.2",
  "seaborn>=0.13.2",
  "torch>=2.6.0",
  "tqdm>=4.67.1",
  "xgboost>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.2",
]

[project.scripts]
bioformer-train-baseline = "bioformer.train.train_baseline:main"
bioformer-train-transformer = "bioformer.train.train_transformer:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

